+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -12,10 +12,17 @@
 @import flutter_ble_peripheral_slave;
 #endif
 
+#if __has_include(<permission_handler_apple/PermissionHandlerPlugin.h>)
+#import <permission_handler_apple/PermissionHandlerPlugin.h>
+#else
+@import permission_handler_apple;
+#endif
+
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
   [BlePeripheralPlugin registerWithRegistrar:[registry registrarForPlugin:@"BlePeripheralPlugin"]];
+  [PermissionHandlerPlugin registerWithRegistrar:[registry registrarForPlugin:@"PermissionHandlerPlugin"]];
 }
 
 @end